server/requirements.txt
Flask==0.11.1
cf-deployment-tracker==1.0.2
cloudant==2.4.0
line-bot-sdk==1.1.0
